Removed 32-bit project files.  All projects will be for 64-bit platforms.
Added library dependencies to release version of projects (CUDA/OpenCV)

To make the sample code legible, the calls to the kernel function using
the shared and global memory accesses were put into their own functions (including
the setup of the kernel function calls), each of which was put in a separate
file: FilterOnGlobalMemory.cpp and FilterOnSharedMemory.cpp.

In the main() function, the calls to the Filter kernel functions on both the
shared memory and relying on global memory access and their setups were replaced
by the wrapper functions FilterOn(Global/Shared)Memory() to keep the logic
of the programme legible.